refactor() reviewing refactoring with actual pipeline runs
