+++ src/core/services/loading/__init__.py
@@ -2,17 +2,12 @@
 Loading service components for data loading operations.
 """
 
-from .memory_manager import MemoryManager
 from .batch_processor import BatchProcessor
-from .file_processor import FileProcessor
+from .file_handler import FileHandler
+from .service import FileLoadingService
 
 __all__ = [
-    'FileLoader',
-    'batch_generator_csv', 
-    'batch_generator_parquet',
-    'async_upsert',
-    'create_pool',
-    'map_types',
-    'ensure_table_sql',
-    'upsert_from_temp_sql'
+    "BatchProcessor",
+    "FileHandler",
+    "FileLoadingService",
 ]
