+++ src/core/services/loading/batch_processor.py
@@ -96,6 +96,8 @@ class BatchProcessor:
         timestamp = datetime.now().strftime("%Y%m%d_%H%M%S_%f")[:-3]
         batch_name = f"MemoryAware_Batch_{table_name}_chunk{batch_num}_{len(batch_chunk)}rows_{timestamp}"
 
+        print('Teste 2')
+
         with self.batch_context(
             table_name=table_name,
             batch_name=batch_name,
