+++ src/setup/config/validation.py
@@ -6,9 +6,9 @@ and relationships between different configuration sections.
 """
 
 from typing import List, Tuple, Optional
-import os
 import logging
-import multiprocessing
+import os
+import re
 
 from .models import AppConfig, Environment
 
@@ -41,30 +41,26 @@ class ConfigurationValidator:
     def _validate_memory_constraints(self):
         """Validate memory usage will not exceed limits."""
         try:
-            conversion = self.config.pipeline.conversion
+            memory_config = self.config.pipeline.memory
+            conversion_config = self.config.pipeline.conversion
             
-            # Estimate memory usage for conversion
+            # Estimate memory usage for conversion (approximate)
             estimated_memory_mb = (
-                conversion.chunk_size * 
-                conversion.workers * 
-                conversion.row_estimation_factor
+                conversion_config.chunk_size * 
+                conversion_config.workers * 
+                conversion_config.row_estimation_factor
             ) / (1024 * 1024)
             
-            if estimated_memory_mb > conversion.memory_limit_mb:
+            if estimated_memory_mb > memory_config.memory_limit_mb:
                 self.errors.append(
-                    f"Estimated memory usage ({estimated_memory_mb:.1f}MB) exceeds "
-                    f"conversion memory limit ({conversion.memory_limit_mb}MB). "
-                    f"Consider reducing chunk_size ({conversion.chunk_size}) or "
-                    f"workers ({conversion.workers})"
+                    f"Estimated conversion memory {estimated_memory_mb:.1f}MB exceeds memory_limit_mb {memory_config.memory_limit_mb}MB"
                 )
             
-            # Warn if memory usage is very high
-            if estimated_memory_mb > conversion.memory_limit_mb * 0.8:
+            # Warn if memory usage is very high (>80% of limit)
+            if estimated_memory_mb > memory_config.memory_limit_mb * 0.8:
                 self.warnings.append(
-                    f"High memory usage ({estimated_memory_mb:.1f}MB) approaching "
-                    f"limit ({conversion.memory_limit_mb}MB)"
+                    f"Estimated conversion memory {estimated_memory_mb:.1f}MB is >80% of memory limit ({memory_config.memory_limit_mb}MB)"
                 )
-                
         except Exception as e:
             self.errors.append(f"Memory validation failed: {str(e)}")
     
@@ -83,31 +79,23 @@ class ConfigurationValidator:
             # Error if total workers significantly exceed available cores
             if total_workers > available_cores * 2:
                 self.errors.append(
-                    f"Total workers ({total_workers}) significantly exceeds "
-                    f"available CPU cores ({available_cores}). This may cause "
-                    f"performance degradation."
+                    f"Total workers ({total_workers}) is > 2x available cores ({available_cores}); risk of oversubscription"
                 )
-            
-            # Warning if workers exceed cores by reasonable margin
-            elif total_workers > available_cores * 1.5:
+            elif total_workers > int(available_cores * 1.5):
                 self.warnings.append(
-                    f"Total workers ({total_workers}) exceeds available "
-                    f"CPU cores ({available_cores}) by significant margin"
+                    f"Total workers ({total_workers}) is >1.5x available cores ({available_cores}); consider reducing parallelism"
                 )
             
             # Validate individual worker counts
-            if pipeline.conversion.workers > available_cores:
+            if pipeline.conversion.workers > max(1, available_cores):
                 self.warnings.append(
-                    f"Conversion workers ({pipeline.conversion.workers}) exceed "
-                    f"available cores ({available_cores})"
+                    f"Conversion workers ({pipeline.conversion.workers}) exceed available cores ({available_cores})"
                 )
             
-            if pipeline.loading.workers > available_cores // 2:
+            if pipeline.loading.workers > max(1, available_cores // 2):
                 self.warnings.append(
-                    f"Loading workers ({pipeline.loading.workers}) may be too high "
-                    f"for database operations"
+                    f"Loading workers ({pipeline.loading.workers}) is large relative to available cores ({available_cores})"
                 )
-                
         except Exception as e:
             self.errors.append(f"Worker validation failed: {str(e)}")
     
@@ -119,38 +107,32 @@ class ConfigurationValidator:
             
             # Loading batch size vs conversion chunk size
             if loading.batch_size > conversion.chunk_size:
-                self.errors.append(
-                    f"Loading batch size ({loading.batch_size}) cannot exceed "
-                    f"conversion chunk size ({conversion.chunk_size})"
+                self.warnings.append(
+                    f"Loading.batch_size ({loading.batch_size}) > conversion.chunk_size ({conversion.chunk_size}); consider aligning these values"
                 )
             
             # Sub-batch size vs batch size
             if loading.sub_batch_size > loading.batch_size:
                 self.errors.append(
-                    f"Loading sub-batch size ({loading.sub_batch_size}) cannot exceed "
-                    f"batch size ({loading.batch_size})"
+                    f"Loading.sub_batch_size ({loading.sub_batch_size}) cannot be greater than loading.batch_size ({loading.batch_size})"
                 )
             
             # Min vs max batch sizes
             if loading.min_batch_size > loading.max_batch_size:
                 self.errors.append(
-                    f"Min batch size ({loading.min_batch_size}) cannot exceed "
-                    f"max batch size ({loading.max_batch_size})"
+                    f"Loading.min_batch_size ({loading.min_batch_size}) is greater than loading.max_batch_size ({loading.max_batch_size})"
                 )
             
             # Batch size vs min/max ranges
             if loading.batch_size < loading.min_batch_size:
-                self.warnings.append(
-                    f"Batch size ({loading.batch_size}) is below minimum "
-                    f"({loading.min_batch_size})"
+                self.errors.append(
+                    f"Loading.batch_size ({loading.batch_size}) is smaller than min_batch_size ({loading.min_batch_size})"
                 )
             
             if loading.batch_size > loading.max_batch_size:
-                self.warnings.append(
-                    f"Batch size ({loading.batch_size}) exceeds maximum "
-                    f"({loading.max_batch_size})"
+                self.errors.append(
+                    f"Loading.batch_size ({loading.batch_size}) is greater than max_batch_size ({loading.max_batch_size})"
                 )
-                
         except Exception as e:
             self.errors.append(f"Batch hierarchy validation failed: {str(e)}")
     
@@ -158,77 +140,59 @@ class ConfigurationValidator:
         """Apply environment-specific validations."""
         try:
             pipeline = self.config.pipeline
+            available_cores = os.cpu_count() or 4
             
             if pipeline.environment == Environment.DEVELOPMENT:
-                # Development-specific warnings
-                if pipeline.conversion.chunk_size > 10000:
-                    self.warnings.append(
-                        f"Large conversion chunk size ({pipeline.conversion.chunk_size}) "
-                        f"in development mode may cause memory issues"
-                    )
-                
-                if pipeline.loading.batch_size > 5000:
-                    self.warnings.append(
-                        f"Large loading batch size ({pipeline.loading.batch_size}) "
-                        f"in development mode may be inefficient"
-                    )
-                
-                if not pipeline.development.enabled:
-                    self.warnings.append(
-                        "Development optimizations are disabled in development environment"
-                    )
+                # Development can be relaxed but should warn on production-like settings
+                if pipeline.loading.batch_size > 100000:
+                    self.warnings.append("Development environment with very large loading.batch_size; consider lowering for dev runs")
             
             elif pipeline.environment == Environment.PRODUCTION:
-                # Production-specific validations
-                if pipeline.development.enabled:
-                    self.errors.append(
-                        "Development mode cannot be enabled in production environment"
-                    )
-                
-                if pipeline.conversion.chunk_size < 10000:
-                    self.warnings.append(
-                        f"Small conversion chunk size ({pipeline.conversion.chunk_size}) "
-                        f"in production may be inefficient"
-                    )
-                
-                if pipeline.loading.batch_size < 1000:
-                    self.warnings.append(
-                        f"Small loading batch size ({pipeline.loading.batch_size}) "
-                        f"in production may be inefficient"
-                    )
+                # Stricter requirements for production
+                if pipeline.conversion.workers > available_cores * 2:
+                    self.errors.append("Production conversion.workers too large relative to available cores")
+                if pipeline.development and getattr(pipeline.development, "enabled", False):
+                    self.warnings.append("Development mode enabled in production environment")
             
             elif pipeline.environment == Environment.TESTING:
-                # Testing-specific validations
-                if pipeline.conversion.chunk_size > 5000:
-                    self.warnings.append(
-                        f"Large conversion chunk size ({pipeline.conversion.chunk_size}) "
-                        f"in testing environment may slow down tests"
-                    )
-                    
+                # Testing should use small batches to speed up tests
+                if pipeline.loading.batch_size > 5000:
+                    self.warnings.append("Testing environment with large loading.batch_size; consider reducing to speed tests")
         except Exception as e:
             self.errors.append(f"Environment-specific validation failed: {str(e)}")
     
     def _validate_database_connectivity(self):
         """Validate database configuration completeness."""
         try:
-            for db_name, db_config in self.config.databases.items():
-                if not db_config.host:
+            # Validate pipeline main DB (attached to data_sink) and audit DB (attached to audit)
+            candidates = {
+                "pipeline_main": getattr(self.config.pipeline.data_sink, "database", None),
+                "audit_db": getattr(self.config.audit, "database", None)
+            }
+            for db_name, db_config in candidates.items():
+                if db_config is None:
+                    self.errors.append(f"Database '{db_name}' is not configured")
+                    continue
+
+                if not getattr(db_config, "host", None):
                     self.errors.append(f"Database '{db_name}' missing host configuration")
-                
-                if not db_config.user:
+
+                if not getattr(db_config, "user", None):
                     self.errors.append(f"Database '{db_name}' missing user configuration")
-                
-                if not db_config.password:
+
+                if not getattr(db_config, "password", None):
                     self.warnings.append(f"Database '{db_name}' missing password configuration")
-                
-                if not db_config.database_name:
+
+                if not getattr(db_config, "database_name", None):
                     self.errors.append(f"Database '{db_name}' missing database name")
-                
-                # Validate port ranges
-                if not (1 <= db_config.port <= 65535):
-                    self.errors.append(
-                        f"Database '{db_name}' port ({db_config.port}) out of valid range"
-                    )
+
+                port = getattr(db_config, "port", None)
+                try:
+                    port_int = int(port)
+                    if not (1 <= port_int <= 65535):
+                        self.errors.append(f"Database '{db_name}' port ({port}) out of valid range")
+                except Exception:
+                    self.errors.append(f"Database '{db_name}' has invalid port: {port}")
                     
         except Exception as e:
             self.errors.append(f"Database validation failed: {str(e)}")
@@ -245,44 +209,47 @@ class ConfigurationValidator:
                 ("conversion", paths.conversion)
             ]:
                 if not path_value:
-                    self.errors.append(f"Path '{path_name}' is empty")
+                    self.errors.append(f"Path '{path_name}' is not configured")
                     continue
-                
-                # Check for absolute paths that might be problematic
-                if os.path.isabs(path_value) and not os.path.exists(os.path.dirname(path_value)):
-                    self.warnings.append(
-                        f"Path '{path_name}' ({path_value}) parent directory does not exist"
-                    )
-                    
+
+                # If path exists ensure it's a directory and writable
+                if os.path.exists(path_value):
+                    if not os.path.isdir(path_value):
+                        self.errors.append(f"Path '{path_name}' ({path_value}) exists but is not a directory")
+                    else:
+                        # Check writability
+                        if not os.access(path_value, os.W_OK):
+                            self.warnings.append(f"Path '{path_name}' ({path_value}) is not writable by the current user")
+                else:
+                    # Not existing is a warning (we often create dirs at runtime)
+                    self.warnings.append(f"Path '{path_name}' ({path_value}) does not exist; it will be created at runtime if needed")
         except Exception as e:
             self.errors.append(f"Path validation failed: {str(e)}")
     
     def _validate_url_connectivity(self):
         """Validate URL configuration format."""
         try:
-            urls = self.config.urls
-            
-            # Basic URL format validation
-            for url_name, url_value in [
-                ("base_url", urls.base_url),
-                ("layout_url", urls.layout_url)
-            ]:
+            # URLs are stored on pipeline.data_source in current loader
+            ds = getattr(self.config.pipeline, "data_source", None)
+            if ds is None:
+                self.errors.append("Data source configuration (URLs) is missing")
+                return
+
+            url_map = {
+                "base_url": getattr(ds, "base_url", None),
+                "layout_url": getattr(ds, "layout_url", None)
+            }
+
+            for url_name, url_value in url_map.items():
                 if not url_value:
                     self.errors.append(f"URL '{url_name}' is empty")
                     continue
-                
-                if not url_value.startswith(('http://', 'https://')):
-                    self.errors.append(
-                        f"URL '{url_name}' ({url_value}) must start with http:// or https://"
-                    )
-                
-                # Check for localhost in production
-                if (self.config.pipeline.environment == Environment.PRODUCTION and 
-                    'localhost' in url_value):
-                    self.warnings.append(
-                        f"URL '{url_name}' contains localhost in production environment"
-                    )
-                    
+
+                if not isinstance(url_value, str) or not url_value.startswith(("http://", "https://")):
+                    self.errors.append(f"URL '{url_name}' ({url_value}) must start with http:// or https://")
+
+                if (self.config.pipeline.environment == Environment.PRODUCTION and "localhost" in url_value):
+                    self.warnings.append(f"URL '{url_name}' contains localhost in production environment")
         except Exception as e:
             self.errors.append(f"URL validation failed: {str(e)}")
     
@@ -300,13 +267,13 @@ class ConfigurationValidator:
         warning_count = len(self.warnings)
         
         if error_count == 0 and warning_count == 0:
-            return "✅ Configuration validation passed with no issues"
+            return "Configuration validation: OK (no errors or warnings)"
         
         summary = []
         if error_count > 0:
-            summary.append(f"❌ {error_count} error(s)")
+            summary.append(f"{error_count} error(s)")
         if warning_count > 0:
-            summary.append(f"⚠️ {warning_count} warning(s)")
+            summary.append(f"{warning_count} warning(s)")
         
         return f"Configuration validation completed: {', '.join(summary)}"
     
@@ -315,17 +282,15 @@ class ConfigurationValidator:
         log = logger or logging.getLogger(__name__)
         
         if self.has_errors():
-            log.error("Configuration validation errors:")
-            for error in self.errors:
-                log.error(f"  - {error}")
+            for e in self.errors:
+                log.error(e)
         
         if self.has_warnings():
-            log.warning("Configuration validation warnings:")
-            for warning in self.warnings:
-                log.warning(f"  - {warning}")
+            for w in self.warnings:
+                log.warning(w)
         
         if not self.has_errors() and not self.has_warnings():
-            log.info("Configuration validation passed successfully")
+            log.info("Configuration validation passed with no issues")
 
 
 def validate_configuration(config: AppConfig) -> Tuple[List[str], List[str]]:
